+++ ctlearn/tools/utils.py
@@ -1,15 +1,27 @@
+"""
+Utility functions for the CTLearn tools.
+"""
+
+from enum import Enum
+import pathlib
 from importlib.resources import files, as_file
+import os
+import time
+from tqdm import tqdm
 
 from ctapipe.core import Provenance
 from ctapipe.core.traits import TraitError
 from ctapipe.instrument.optics import FocalLengthKind
 from ctapipe.instrument import SubarrayDescription
 
-import os
-import time
-from tqdm import tqdm
 
-__all__ = ["validate_trait_dict","get_lst1_subarray_description","monitor_progress"]
+__all__ = [
+    "monitor_progress",
+    "validate_trait_dict",
+    "get_lst1_subarray_description",
+    "FrameworkType",
+    "detect_framework",
+]
 
 def monitor_progress(src_path, dst_path, stop_event, logger):
     try:
@@ -86,3 +98,52 @@ def get_lst1_subarray_description(focal_length_choice=FocalLengthKind.EFFECTIVE)
     with as_file(files("ctlearn") / "resources/LST-1_SubarrayDescription.h5") as path:
         Provenance().add_input_file(path, role="SubarrayDescription")
         return SubarrayDescription.from_hdf(path, focal_length_choice=focal_length_choice)
+
+class FrameworkType(Enum):
+    """
+    Deep learning framework type enumeration.
+    
+    This enumeration specifies which deep learning framework to use for
+    model training and inference. CTLearn supports both Keras (TensorFlow backend)
+    and PyTorch frameworks.
+    
+    Attributes:
+        KERAS (int): Use Keras/TensorFlow framework (value: 1)
+            - Advantages: High-level API, easy to use, good for prototyping
+            - TensorFlow 2.x with Keras API
+            - Suitable for production deployment
+            
+        PYTORCH (int): Use PyTorch framework (value: 2)
+            - Advantages: Dynamic computation graphs, flexible, research-friendly
+            - PyTorch 1.x or 2.x
+            - Better for custom architectures and experimental models
+    
+    Example:
+        >>> from ctlearn.core.ctlearn_enum import FrameworkType
+        >>> framework = FrameworkType.PYTORCH
+        >>> print(framework.name)  # 'PYTORCH'
+        >>> print(framework.value)  # 'PyTorch'
+    """
+    KERAS = "Keras"
+    PYTORCH = "PyTorch"
+
+def detect_framework(path_val):
+    """
+    Determines framework based on file extension.
+    Returns 'Keras', 'PyTorch', or raises TraitError.
+    """
+    if path_val is None:
+        return None
+
+    path = pathlib.Path(path_val)
+    ext = path.suffix.lower()
+
+    if ext in [".keras", ".h5"]:
+        return FrameworkType["KERAS"]
+    elif ext in [".pt", ".pth"]:
+        return FrameworkType["PYTORCH"]
+    else:
+        raise TraitError(
+            f"Invalid model extension '{ext}' for file '{path}'. "
+            "Expected '.keras' or '.h5' for Keras, or '.pt' or '.pth' for PyTorch."
+        )
